Add Applicant#mode
Why:

We want applicants with transit passes to have the option to use them to
get to their jobs, so when we look up travel times, we need to do it by
mode.

This change addresses the need by:

* Adding a 'mode' method to Applicant which uses the has_transit_pass
boolean to determine what mode should be looked up to get the travel
time, walking or transit.