Define scope for available positions.
Why:

When applicants are placed in a position, that position becomes
unavailable, but only for that run. A position is considered available
if it hasn't been placed in that run.

This change addresses the need by:

* Scoping available placements, and testing this.

NOTE: The test took a lot of external collaborators to set up. I'd like
to find a way to run a test like this without requiring so much database
setup.