Add status and expiration to Placement.
Why:

In order to correctly relay acceptance or declining to ICIMS, we'll have
to check to make sure the Placement isn't expired, and we'll set the
expiration manually when DYEE sends their mass email.

We'll also need to mark placements as accepted, declined, expired, or,
if the person has already been hired in ICIMS, as 'invalid'.

This change addresses the need by:

* Adding an expiration date and an instance method #expired? to
Placement.

* Adds a status to placement that defaults to :potential, and can be set
to :accepted, :declined, or :invalid.