Consolidate databases.
Why:

Databases were getting confused, trying to establish multiple
connections. It also made Grid and TravelTime, and anything depending on
those, untestable -- we have a DatabaseCleaner, which could have wiped
out tables were we connecting to db.prep in the test environment.

This change addresses the need by:

* Creating migrations for the test and development environments. We do
not need to run these migrations in production.

* Removing the secondary database connection from the configuration.

* Changing per-model connections to use the primary database, with the
secondary being gone.

* Moving the enabling of PostGIS to its own migration.

* Bonus: Adding ERB potential to the database configuration. We're not
using it at the moment -- DATABASE_ENV instead -- but it's nice to be
able to use that if we want it later.