Clean up tests.
Why:

As an artifact of trying to connect to a populated database, and
conflating fixtures with importable data, the tests are a little
confused. In order to improve the architecture and ensure stability,
making the tests simple, fast, and isolated.

This change addresses the need by:

* Replacing the fixture based strategy `Model.first` with setup and
teardown methods. This will help ensure the test database doesn't get
polluted with data.

* Removing assert_send. It's unclear what it's trying to test. We'll
replace these tests in the next commit.