Test #within query.
Why:

The #within query has a lot going on -- want to make sure it's tested
well to avoid regression.

This change addresses the need by:

* Setting up a basic test case for the query.

* Removing #within?, since, at the moment, we don't have a need for it.
When we do, we can find a clearer way to write it.

NOTE: At the moment, there's a lot going on in #within, and there's
likely a better way to write it. In the interest of getting things
shipped, however, we'll leave it how it is for now.