Add GeoJSON export to StatsJob
Why:

In order to explore the data visually, we want GeoJSON to map. We need
to see applicants and be able to differentiate placed from unplaced
applicants visually. We'd also like to see the placements (as
straight-line commutes) and the scores / mode associated with
each placement.

This change addresses the need by:

* Edited StatsJob to put out a simple GeoJSON FeatureCollection.

* Wired rake task to use StatsJob.

* Commented out parts of StatsJob that I'm not working on.

NOTE: Maybe StatsJob isn't the right place for this. Or maybe it is, but
we should store the GeoJSON and export it later instead. Should we have
a task match:post that generates statistics, exports a CSV, and exports
the GeoJSON for a run?