Export GeoJSON, and chain stats & export to match.
Why:

We need a way to export GeoJSON, so we're doing it with the placement
CSV for the moment. Also, generating stats and exporting takes too many
commands, so we need to chain it.

This change addresses the need by:

* Returning a Run ID in the Match job, and using it to invoke the stats
generation and export tasks.

* Adding GeoJSON export to CSV export.

NOTE: Need a way to see the core stats.