Debug deterministic matching
Why:

Every run was placing applicants into the same jobs each time, leading
to the conclusion that the randomization was being overridden somewhere.
We discovered that #find_each silently overrides any order or offset on
the scope, which meant that our random order was being overridden to
become `order(:id)`.

We explored several solutions, but batch-querying records randomly
appears to be a fairly difficult challenge. So, @allthesignals proposed
we just pluck the IDs, since 6000 IDs is within our memory limits,
whereas 6000 ActiveRecord instances might not be.

This change addresses the need by:

* Plucking ID from the randomized database records, and using that ID to
look up each applicant in turn.

* Add descriptive_statistics, which got lost in a botched merge.