Start a web app for examining runs.
Why:

We had gotten to a point where viewing statistics and copying GeoJSON
into another map was becoming cumbersome and annoying. Instead of
depending on rake for output, why not build a quick Sinatra app?

Additionally, when OpenStructs are directly serialized to JSON, they
come with a 'table' component and a 'modifiable' component, which
adding a key to data access.

This change addresses the need by:

* Merging an app file and Gemfile from relay, but deleting all of the
content, because we don't need to let relay code get in our way.

* Adding rerun, to reload the app as it changes. (Too bad there's no
zero-downtime rerun.)

* Converting the statistics OpenStruct to a hash before serializing it
to JSON, thus avoiding the extra ostruct-specific keys.