Finish up refactor
Why:

After the last commit, we still needed to add the new fields to
placement to track opportunities. After doing so, there were a lot of
nil positions, and had to update the stats task to handle those nils.

This change addresses the need by:

* Add opportunity scores and an applicant's position in the lottery to
the placement model.

* Debug export task, guarding against nils and ensuring the output of
valid GeoJSON.

This commit introduces a modicum of technical debt through a lack of
tests. A lot of the introduced functionality could and should be tested.