Fix histogram; add stats, delete to index page
Why:

In order to compare runs, we want to add statistics to the runs/index
page.

And in order to run #fetch, etc. on statistics, it can't be null, and
should default to an empty hash.

Also, we found the histogram graph on the runs/show page had hardcoded
values.

This change addresses the need by:

* Rewriting runs/index as a table with statistics and statuses

* Adding a button and route to delete runs

* Migrating statistics to be non-null, and default to an empty hash.

* Fixed histogram display to use actual values.