Add limit, finish stats, rename match -> lottery
Why:

It came up in discussion that we might want to limit how many applicants
are placed at once, so as to produce more manageable batches.

Additional

Giving the rake tasks a namespace of 'match' did not make sense when
demonstrating the commands. We're running the lottery, so it should be
'lottery:run'.

This change addresses the need by:

* Adding an optional limit argument to lottery:run.

* Finishing the statistics calculations.

* Renaming 'match' tasks to 'lottery' tasks.