Add ICIMS API interaction
Why:

In order to interact with the ICIMS API in a controlled, structured way,
it made sense to create a module -- which would be easily extracted as a
gem -- to interact with ICIMS.

This change addresses the need by:

* Adding a base ICIMS resource that sets the base URL and headers for
the API, and handles response errors.

* Add an address module (not a resource) that takes in ICIMS-formatted
'addresses' objects and turns them into simple string addresses, for
geocoding.

* Add a company resource, to refer to when getting the location of a
given job.

* Add a job resource that references the number of remaining positions.

* Add a person resource, which gets all of the attributes we need for
lottery placement, including transit pass, location/interest preference,
interests, and geocodable address.

* Add a workflow resource, so we can see what jobs the applicant has
applied to or is connected to. Eventually, this resource will need to
create and update workflows -- the equivalent of placing someone in a
job.

* We add test fixtures, but don't do anything with them yet. It would be
wise to test these objects using fixtures, so we don't have to
bin/console every time we want to make sure a response is being parsed
correctly.