Add scopes for eligible jobs, workflows.
Why:

Need to know who is eligible to be imported into the database in order
to import them -- don't want to bring in everybody.

Also, you can't determine eligible people, only workflows, with people
attached. For our purposes, it seems as if a person's status is really
the combination of the statuses of all of their workflows.

This change addresses the need by:

* Adding scope to request eligible records.

* Adds :limit option to only iterate through some records.

* Add test fixtures for responses.

NOTE: Need an :offset option, in case the process crashes halfway
through, or if we want to page through. Maybe add Kaminari to paginate
an array.