Add Workflow creation, updating
Why:

To place a person into a job, then indicate that they've either accepted
or declined, we need methods to create and update workflows.

This change addresses the need by:

* Add .create, #save, #update to POST and PATCH the Workflow resource.

* Mimic Rails-style new/create/save returns.

* Add state-helper methods to make accepting, declining, placing logic
simpler.

NOTE: Should check before updating status to see if the workflow is
available to be placed in a certain workflow status, and fail in a
particular way if not.

* Move Content-Type into main headers, so we don't have to merge that in
with every request.

* Make attributes an indifferent hash.