Ensure workflow is updatable before updating
Why:

We don't want to update people who are hired, CBO submitted, or another
status where we shouldn't be allowed to update the workflow status.

This change addresses the need by:

* Adding a check to make sure a workflow is in Lottery:Accepted,
Lottery:Declined, Lottery:Placed, or a New Submission before updating
the workflow status. We can still go back and revert between accepted /
declined, but we can't mess with workflows that are, for example, hired.