Add category conversion, prepare Relay tests.
Why:

We need to make sure that the right actions occur in the Relay app, and
that certain conditions are checked for before accepting or declining a
placement in the hiring system.

In order to check interests matching, and to get around an impasse with
an external service, we need to split interests into an array of
categories, and see if there is any intersection.

Tests weren't failing on InterestScore because we had mocked
:category to return a static value. We either needed an integration
test - which should be coming in the next commit - or should have used a
fixture.

This change addresses the need by:

* Add CategorySplitter

* Alter tests on jobs, interest score to use array categories.

NOTE: Still need to migrate the database.

* Set up test scaffolding for Apps::Relay. We add Rack::Test to check
responses, and set up all the possible cases we can think of for the
relay to run into.

* Add failing tests to test conversion between ICIMS and local
resources.

* Update gems.