Add constructors to convert from ICIMS
Why:

We need to be able to import data from ICIMS to our local environment
very easily.

This change addresses the need by:

* Limit the attributes sent to the constructor to the canonical database
attributes.

* Process interests early on in the constructor, and ensure the
ICIMS resource attributes needed for the local resources are available
as instance variables, since that's how the attributes hash is decided.

* Make the validation of grid_id conditional, and dependent on their
being a location. We're not geocoding as we're importing.