Start testing Relay app
Why:

To develop the Relay web app without constantly pinging external
services, we are test-driving the app with mocks.

This change addresses the need by:

* Adds initial tests to Relay app.

* Adds status attribute to local Placement object.

NOTE: While we want to permanently redirect people (308) who accept or
decline, we can't with this version of Rack and still test. The master
branch considers 308 a redirect, but pulling in master breaks too many
other pieces.