Sketch better placement design
Why:

Starting to build in the equity component, altering the way jobs are
allocated, scored, selected.

This change addresses the need by:

* Changing the way we think about available jobs from JobFinder to
(Job)Pool, which gives us an clearer object to think about.

* Adding ScoredPosition, to attach a score to a position, to help find
the best fit from the Pool. Not sure this is the right approach.

* Add Compressor, the equity component. Thinking about job allocation
like signal strength: Opportunity is the signal. Lower opportunity needs
a boost to get to an acceptable range; higher opportunity is left alone.