Add Pool, PooledPosition resources.
Why:

To better control the way that data about placement and positions is
stored, and the way jobs are found and opportunities are balanced, we

This change addresses the need by:

* Adds Pool, which is a collection of PooledPositions.

* Add PooledPositions, which is associated with an Applicant and Run
(through a pool) and is given a score.

* Add basic tests for attributes that Pool should save. NOTE: These need
more work.

* Add scopes that make finding jobs easier. NOTE: These need to be
implemented elsewhere.

* Change grid_id validation Position to be conditional on
having a point location.

* Add :reserve attribute to Position, for citywide positions, aka those
used in balancing / compression.

* Alter score JSON to not include applicant and position.