Connect Placement and Workflow
Why:

We'd like to interact ONLY with Placement in the Relay app, which means
that Placement needs to own ICIMS::Workflow, and synchronize with it.

This change addresses the need by:

* Adding workflow_id to Placement, to keep track of which workflow the
placement is connected to.

* Skip the tests for placement/decline, and just focus on
placement/accept. When accept is ready, so will decline. To do both at
once is double-work.

* Check if the placement has already been accepted or declined locally,
then check with the remote resource.

* Add a 'placed' status on Placement, for placements that have been
'finalized' and are ready to be accepted or declined. The status
'pending' is for placements that have not been finalized.

* Add an error case to catch 500s. I don't think it does.

NOTE: There is a significant failing test due to Placement and
ICIMS::Workflow not being fully connected yet.

NOTE: Finding the need for a status conversion object.