Align placement and workflow decision tests
Why:

We want to check a number of preconditions before placing, accepting, or
declining a placement.

This change addresses the need by:

* Update null object to mimic ICIMS::Workflow, adding methods to
implicitly check presence (if we need to later).

* Update tests to place before deciding, since unplaced workflows cannot
be decided upon.

NOTE: The workflow/placement tests seem a little intermeshed, because
there's a decent amount of responsibility-sharing. If there's free time
to refactor later, it might be good to check in on this.