Check expiration.
Why:

We need to know if the offer is expired.

This change addresses the need by:

* Adding an expires_at timestamp to placements. We don't set it yet --
that will come in the next commit.

* Check to see if the status is set to expired. If not, check the
expires_at time against the current time, and set the status to expired
if it is.

NOTE: You can't revert to unexpired just by setting a different
expiration date. It's still expired, because the status has been set.
This might be good to validate, so people aren't reverting dates
thinking they're unexpired.