Validate that the expiration date is in the past
Why:

It's possible we'll make a mistake in setting expiration dates, and want
to go back and set them to be in the future, so applicants have more
time.

However, the status would still be 'expired', because once the record
determines that it's expired, it sets a status of 'expired', and never
knows to revert it.

This change addresses the need by:

* Adding a custom validation that ensures that, if the status is still
'expires', the date is in the past. If someone were to change the status
AND the expiration date, this validation would pass.