Restructure environment, mostly for Heroku
Why:

Loading files in Heroku was error prone: unlike the local environment,
the load order mattered.

The environment file -- which loads all of the project into the
console / web app -- was a chunk of unstructured Ruby script. In the
spirit of the rest of this project, we want to rearchitect it to be
modular, so we can make changes as needed.

This need arose partly because the database configuration was getting
set differently in different places, and we don't want to deal with
that.

This change addresses the need by:

* We add explicit load order, requiring as needed at the top of each
file.

* Make the database URL a string in production, hashes in development in
test.

* Wrapping all pre-file-load actions in an Initializer module, with
submodules taking care of loading the configuration file(s), connecting
to the database, and setting up a logger.

> We're kind of re-inventing Rails initializers here.

> At some point, we'll likely have a lottery configuration file. We'll
load that through an initializer as well.

* Setting up the Initializers top module to load all of its submodules.
Each submodule must implement #load.

* After all this setup, delegating the loading of configuration in
the `rake db` suite to the initializers. This should keep things
consistent.

* Tried to make the Initializers::Database module as standalone as
possible, loading the config it needs if it errors the first time. Not
sure why I did this.