Add compressor and configuration.
Why:

In order to give applicants living in low-opportunity areas a fair
chance at getting placed in a job, we add upward signal compression --
the signal being "opportunity".

We want to be able to tweak the compression to get it just right, and we
want to know which sample runs use which compression setting.

This change addresses the need by:

* Adding a compressor that reads the lottery.yml configuration.

* Add simple compressor configuration to the lottery.yml file.

* Store the lottery configuration settings in each run, so we can go
back to see what the configuration was for a particular run.

* Move seed-setting from MatchJob into the Run itself -- MatchJob was
micromanaging Run.