Set up preparation tasks.
Why:

Before running the lottery, we need to assign random ticket / order
numbers to applicants, and precalculate the base pools
(pre-compression).

Additionally, pool relationships were becoming very convoluted.
Placement emerged as the core model, connecting runs, applicants, and
positions.

This change addresses the need by:

* Adding ticket assignment job.

* Simplifying pool relationships, attaching them to placements instead
of the combination of runs and applicants (which can be simplified to
placements)