Finish up best_fit, remove junk methods
Why:

We hadn't quite gotten a handle on best_fit, but managed to figure it
out without adding the extra data structures originally planned.

Refactoring left behind some unnecessary, messy methods.

This change addresses the need by:

* Selecting best_fit, comparing the number of placements for a given job
with the number of positions for that job.

* Removing the unnecessary methods.

NOTE: There have been a lot of test-free commits. It's probably time to
start putting tests back in line with the code.