Compress based on jobs available at runtime
Why:

Previously, we were basing compression off of the entire base pool,
which is precalculated before the run starts. We want to base
compression off of opportunity at runtime, that is, the number of
opportunities someone has at the moment they are being considered for
placement.

This change addresses the need by:

* Adding an #available? helper that checks whether a position is
presently available within a run.

* Only pulling in the available jobs as the signal to compress.

* Mocking positions for the compressor test, now that we actually have
to call something on the positions themselves.