Add position importer, update applicant importer
Why:

We need to pull in positions as well as applicants, so we add a
position importer to call in a rake task.

Add 'import' rake tasks that pull in data from our external service.
Because the arguments given to rake tasks (in this case: limit,
offset, continue) are sent as strings, we need to convert them before
using them.

This change addresses the need by:

* Adding 'import' rake tasks.

* Creating PositionImporter that largely mirrors ApplicantImporter.

* Adding wannabe_bool to convert string args to booleans.