Add statuses, and fix bugs in ICIMS support libs
Why:

Now that ICIMS added additional statuses for the youth jobs lottery, and
we've added them to our library, they were starting to fill up the file.
We also wanted to add a lookup to get the status symbol/name from the
ICIMS status code. We don't implement this anywhere (yet), but we'll
likely need it soon.

As we started importing addresses, we found some bugs -- sometimes
'addresses' is empty or absent. Also, some applicants (somehow) don't
include the state.

The CategorySplitter missed one category as it was importing positions,
'Hospitality & Tourism', because we expected an '&amp;' and not a
plain '&'.

This change addresses the need by:

* Compacting ICIMS statuses in a hash and defining methods based on the
hash.

* Adding ICIMS status lookup, so we can get a symbol from the ICIMS
 string code.

* In address, arrayify input. Chain #fetch calls to get values that may
or may not be present.

* Adding '&' to the tokens that CategorySplitter can split on.