Fix applicant creation from ICIMS resource
Why:

A test which asked whether an Applicant created via .create_from_icims
was equal to a sample applicant. That applicant even had the wrong
attributes, but it was considered equal for some reason -- and I turned
off warnings, which might have warned me about checking for equality.

So, this was bug-laden.

This change addresses the need by:

* Changed equality test to test equality of attributes, not equality of
resources.

* Add instance variables -- which is what the ICIMS::Person uses
for the #attributes method, and attributes are what the
Applicant.create_from_icims method uses to create the resource.
Basically, by adding instance variables, we are giving #attributes more
attributes, and Applicant can consume those.

* Remove status from ICIMS::Person#find, because it is unclear, even
after much thought, of how to get the status.

* Switching some fixture data around, for tests.