Clarify Positions for importing
Why:

We're splitting categories for interest matching, but that's because
someone during migration failed to align applicant interests with
position categories. We want to keep the original category string, so we
can stratify positions into the automatic or manual market, keeping an
equal proportion of each category in each market.

We wanted to remove duplicate code in handling responses, and only
rescuing certain errors silently.

We needed to check expiration before updating a placement, and
opting-out applicants if they have an expired placement.

This change addresses the need by:

* Add category string back to Position model.

* Use chained #fetch calls to get status (which may be missing). Missing
a person ID or job ID will still raise. We use #fetch instead of #dig
because of our Ruby version, and because implementing dig just for a
two-deep nest will take too long.

* Test status-free workflow responses.

* DRY up response handling, and create a custom error class to raise
ICIMS network errors.