Break ICIMS::Address into components
Why:

Geocoding is less accurate when the address is lumped together into a
string. Also, we want to stratify our sample by zip code, to ensure
relative balance of applicants and positions in each zip code.

This change addresses the need by:

* Breaking address into street, city, state, and zip code.

* Add a helper to convert nine-digit zip codes into five-digit zip
codes; we don't want to stratify by anything smaller than the 5-digit
zip.

* Changing address component accessors from #[] to #fetch, so we error
out if we're missing critical information. We can figure out how to
handle those errors when we import.

* Add to_h helper, so we can more easily represent the address as
a hash of attributes.

* Wire the address object into other ICIMS models.