Convert categories instead of splitting
Why:

Splitting categories into keywords was about to cause a lot of problems,
especially regarding stratification and experimental design. While
talking with our MIT researcher grad student Michael, we decided that
instead of splitting interests into keywords for matching, we should
convert applicant interests into job categories.

Because the lists are out of sync for this year's applications, we have
to do a one-way conversion from applicant interest to job categories --
it's not a one-to-one matching.

This change addresses the need by:

* Adding a CategoryConverter that does one-way conversion from applicant
interests to job categories.

* Removing CategorySplitter and adding CategoryConverter.