Add market allocation to applicants, positions
Why:

After attempting to randomize and stratify entire jobs, we found that it
was unworkable from a statistical / experimental standpoint. We realized
that we needed to be able to place positions at the same job into
different markets.

This change addresses the need by:

* Adding database migrations to accommodate market placement
information, including the allocation of positions in an individual job
to different markets. There's legacy 'market' attribute on position
leftover which hasn't been removed as of this commit.

* Testing to make sure positions have the correct number of allocations.
If a Position object has any open positions (confusing naming)
allocated to either market, the sum of market-allocated positions must
equal the total number of positions in that Position object.