Refactor and rename preparation tasks.
Why:

We have to match the recent changes in how we're thinking about
allocating applicants and positions into markets, randomizing, etc. with
code.

This change addresses the need by:

* Adding an AllocatePositionsJob, which takes in a spreadsheet from our
researchers to assign positions to different markets.

* Adding market allocation into the randomization task. That is, when we
prepare a run through the AssignTicketsJob, we are assigning applicants
(via placements) with a random order and a random market.

* Adding stats so we know how many people got placed into which market,
to ensure an even balance.

* Add more checks to CheckJob, including that positions and placements
are all in markets, and that applicants and positions all have grid ids.

* Ensuring that the precalculation of pools only happens when we don't
have any precalculated pools in the run.