Debug connection between pool, position
Why:

We were finding all manner of strange behavior when working with pooled
positions and compression, including double- or half-counting of
associations (unclear which), compression that didn't build
associations, and more.

This change addresses the need by:

* Fixing the pooled_positions association on Pool.

* Refactoring Pool#compress! to delegate to Compressor#compress!. It's
not a foolproof object design -- Compressor probably knows too much
about Pool and PooledPosition -- but it works.

* Moving the base pool allocation and counting to an after_create hook,
because we need the pool to exist to assign things to it.

* Ensuring a pooled position cannot be entered into a pool twice.

* Changing the #available? method to reference the number of automatic
placements, not the total number of positions.