Update tasks with import, export,
Why:

The LotteryRunJob was using find_each -- which orders by ID no matter
what :order scope is on the query.

The PrecalculatePoolJob was using all placements in a run, not just the
automatic placements -- we don't need to precalculate pools for
everyone.

We needed to use a spreadsheet to start importing applicants, so needed
a task for that.

When exporting mail merge spreadsheets, there's a limited number of
placements we want to export at a time.

Also need to export manual market applicants and positions, so staff can
check if callers are in the manual pool, or if positions have manual
openings left, before placing.

This change addresses the need by:

* Alter LotteryRunJob query to order by index, PrecalculatePool job to
only precalculate auto placements.

* Adding spreadsheet import/export jobs.

* Creating a relevant_placements query for mail merge export, only
getting pending/placed placement data (not accepted/declined).