Debug TravelTime
Why:

We realized that TravelTimes sometimes cannot be found, especially
during compression, when the positions being added to the pool and
scored aren't within the 40-minute travelshed. So, those travel times
are outside the scope of what's in the database, and return a
NoMethodError when we try to get the time in seconds.

Additionally, we want to set a default travel time for those cases. We
want that somewhat configurable, and so want to add it to lottery
config.

We want to ensure that the default behavior is to order by the random
lottery ticket number and not ID, which would remove the point of a
random lottery entirely.

Lastly, found nil-related errors when finding a best_fit position from a
pool of 0 available positions, and needed to rescue that error.

This change addresses the need by:

* Rescue TravelTime error, logging the error and returning the default
travel time as set in config/lottery.yml.

* Add a default_scope to placement -- ordering by index.

* Perform a nil check before returning a position for placement.

* Remove JobFinder, since that's clearly taken over by Pool.