Significant refactor of lottery run.
Why:

This commit represents a massive amount of interconnected changes.

In short, we needed to ensure that only the automatic market was being
chosen for runs, and that we were ordering placements by index, not ID
-- because then we wouldn't be running the lottery in the right order.

Next, the lottery run process was very messy, and while it remains
somewhat messy, we are moving a lot of responsibility for information
about the run, appropriately, into Run.

This change addresses the need by:

* Adds index to applicants; we don't randomize on each run.

* Adds 'synced' status, so we know when someone is placed locally vs
when they are placed via ICIMS.

* Has the placement place itself, based on its pool.

* Log when there's no available best fit, but don't raise on it. It's
going to happen.

* When placements are destroyed, destroy pools and pooled positions.
(We already had: when runs are destroyed, destroy placements.)

* Delegate a lot more responsibility to the run. Each run now knows
which applicants are eligible for it, and which of its placements are
actionable. We are returning an array of IDs presently, but should yield
the object itself.

* Define a method to reload a Run's config, in case we want to run
through the precalculated pools

* Turn the maximum pool size on Compressor into a global. This is
probably a terrible idea, because we have to reload the environment to
unset it. But, it's better than running a 5- to 10-second calculation
every time the compressor runs.

* Default to ordering placement by index.

* Add scopes for automatic market, ordering by index, in every place we
can -- even if it's redundant.

* Start calculating some placement rates on Run. Applicant placement is
not alone a good number, because we won't place every applicant: there
aren't enough jobs. Position fill rate is good, and we should check the
placement rates for the bottom 10% by pool size.