Split interests (workaround)
Why:

ICIMS, in their export, can't choose what delimiter separates elements
in a collection field. So, when they export interests, they are
separated by commas. But because the interests themselves have commas in
them, we can't just use a comma as a token to split on.

So, we have to code a workaround that already knows what possible
interests could be in a string, and breaks them out.

This change addresses the need by:

* Adding and testing an interest splitter.