Maximum pool size
Why:

Tests were showing a nil error during division in the compressor.

This change addresses the need by:

* Converting to integer, so that during tests when there's nil
positions, it defaults to 0.

* Setting max_pool_size directly in tests, because it's a global
variable and no longer stubbable in the same way.

NOTE: It's not great that we're using a global, but that's what we have
for now. We could -- and probably should -- make compressor a singleton.