Swap out HTTP libraries
Why:

In order to interact with ICIMS, we need fixed IP addresses, which we
establish through a proxy service called Fixie.

In order to proxy the HTTPS connection, and maintain port 80 (which is
the port Fixie runs regardless of SSL, since normally it's 443), we need
to switch HTTP libraries. HTTParty was erroring in this case, whereas
Typhoeus was not.

This change addresses the need by:

* Adding Typhoeus as the HTTP library, removing HTTParty.

* Despite webmock saying there's only mocking support for
Typhoeus::Hydra, the only part we needed to change in mocks was stubbing
timeouts, and adding a space to the value in the Authorization header
-- seems like HTTParty was stripping spaces before sending.

* Changing the base URL customer ID and proxy URL information to
environment variables. The ICIMS API URL / customer ID isn't SUPER
sensitive, so we're not going back to scrub the codebase of it, but we
don't need it in there. The proxy URL, however, IS sensitive.