Applicants can't get placed at declined jobs
Why:

If we offer a job to an applicant, and they decline it, we don't want to
offer the same job to them in a second round. There's a pretty good
chance that this will happen, without intervention, since they'd go
first in the second round and that best-fit job would be available since
they declined it.

Also found that we weren't checking job availability based on the right
placements. We were taking all of them, but we really need to ensure
that only want to count placed, synced, and accepted.

  Placed, because we reference that when running the algorithm.
  Synced, because these are in the hiring system but not decided upon
  yet. It's unlikely we'll ever have this status present in the database
  when we run the lottery, but a good safeguard.
  Accepted, because declined and expired positions don't count.

This change addresses the need by:

* Add a #declined_placements helper on Applicant, so we can pull all the
placements they declined within a run.

* Reference the helper in the placement pool, and remove those positions
from the pool.

* Add status query for checking whether a placement is available.