Reorder environment load, and remove ICIMS ID
Why:

Airbrake and CodeClimate were throwing errors in the Travis environment,
and it turned out it wasn't throwing those errors locally because Dotenv
was loading the local .env no matter what. That meant that the actual
Airbrake information was being used locally, instead of a test id and
API key.

We don't want to store the ICIMS ID in the codebase, but it's also not
so critical that we're going back and erasing it.

Tests of the Relay app were erroring when we gave it faulty IDs, because
the applicant and positions lookups were returning nil and you can't
call #id on nil.

This change addresses the need by:

* Conditionally loading .env, not loading it when in the test env.

* Adding the environment name to the console script.

* Handling record not found-related errors for the applicant and
position lookups.

* Changing the ICIMS ID in test data to a random id.