Add to jobs for support
Why:

Before running the lottery, we want to make sure that nobody is missing
critical attributes. We also need to to ensure placements that should be
expired are in fact expired before the next run.

We also needed to import, export, and geocode data more frequently now
that we've switched to spreadsheet-based and not API-based.

This change addresses the need by:

* Raising an error at the end of checking the lottery if any applicant
is missing any critical attribute, and adding a job to check for
expiration.

* Changes the method call from #finalize! to #sync!, to match the change
to the public interface made a while back.

* Adds help in importing, exporting, and geocoding.

* Throughout, debugging and delegating more queries to Run.