Store available positions on Run, NoSQL-style
Why:

The SQL query to get available positions is complex enough that we
started using .find_by_sql, which returns an array and not an
ActiveRecord Relation. This means we can't chain more relations onto it,
which results in problems like "not being able to find the jobs within
a 40-minute radius of an applicant".

So, we store the available positions as an array field on Run. We start
by adding all of the positions. When a placement is updated, it tells
Position to check how many placements have been made with that position,
how that number compares to the number of open positions on
that Position allocated to the automatic market, and to add or remove
itself from the list of positions available in a run, based on that
comparison.

This change addresses the need by:

* Adding an after_save hook on Placement

* Adding a method on Position to check whether it's still available.

* Add a #available_positions integer array, to track IDs. Runs start
with all the positions, then remove/add as necessary. We ensure the
position IDs are unique.